% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabaser.R
\name{build_url}
\alias{build_url}
\title{Assemble URL}
\usage{
build_url(handle, path)
}
\arguments{
\item{handle}{metabase_handle object}

\item{path}{API endpoint path}
}
\description{
Assemble URL
}
\keyword{internal}
