% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabaser.R
\name{metabase_cache_clear}
\alias{metabase_cache_clear}
\title{Clears all the memmory Cache from the metabaser}
\usage{
metabase_cache_clear()
}
\description{
Clears all the memmory Cache from the metabaser
}
