% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabaser.R
\name{metabase_db}
\alias{metabase_db}
\title{Get/Set Metabase DB ID}
\usage{
metabase_db(db_id = NULL)
}
\description{
If no argument is given, retrieves the current Metabase database ID.
If an argument is given, sets the Metabase database ID.
}
