% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabaser.R
\name{metabase_get_default}
\alias{metabase_get_default}
\title{gets a default to be used for metabaser}
\usage{
metabase_get_default(param)
}
\arguments{
\item{param}{normally one of base_url, database_id, creds_file, username, password}
}
\description{
gets a default to be used for metabaser
}
