% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabaser.R
\name{metabase_handle}
\alias{metabase_handle}
\alias{is_metabase_handle}
\title{Construct a Metabase API Handle}
\usage{
metabase_handle(base_url, database_id, username)

is_metabase_handle(handle)
}
\arguments{
\item{base_url}{Base URL for the Metabase API}

\item{database_id}{Database ID to connect to}

\item{username}{Username}
}
\description{
This is constructor to make a metabase_handle object which
is used in subsequent functions to interact with the API.
The handle should not be created manually, but should be created
using \code{\link{metabase_login}}.
}
\section{Functions}{
\itemize{
\item \code{is_metabase_handle}: Check for metabase_handle
}}

\keyword{internal}
