% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabaser.R
\name{metabase_login}
\alias{metabase_login}
\title{Login to Metabase}
\usage{
metabase_login(base_url, database_id, creds_file, auto_setup = TRUE)
}
\arguments{
\item{base_url}{Base URL for the Metabase API}

\item{database_id}{Database ID to connect to}

\item{creds_file}{File containing Metabase account credentials to connect with}

\item{auto_setup}{If TRUE, will automatically setup the connection before login,
if FALSE, requires \code{\link{metabase_setup}} to be executed before}
}
\description{
This will login to the Metabase API with the username and password specified in the given file.
The file should have "username=username" on one line and "password=password" on the next.
If the login is successful, a cookie will be set behind the scense with a session ID for all future requests.
If \code{\link{metabase_setup}} was used, \code{metabase_login} will be automatically configured with a creds_file.
}
