% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabaser.R
\name{metabase_login_default}
\alias{metabase_login_default}
\title{Login to Metabase using default credentials}
\usage{
metabase_login_default(
  base_url = metabase_get_default("base_url"),
  database_id = metabase_get_default("database_id"),
  creds_file = metabase_get_default("creds_file"),
  username = metabase_get_default("username"),
  password = metabase_get_default("password"),
  ...
)
}
\description{
see metabase_login for description of parameters
}
