% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabaser.R
\name{metabase_logout}
\alias{metabase_logout}
\title{Logout of Metabase}
\usage{
metabase_logout(handle)
}
\arguments{
\item{handle}{metabase_handle object}
}
\description{
Logs out the user by sending a request to delete the user session.
}
