% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabaser.R
\name{metabase_query2}
\alias{metabase_query2}
\title{Query Metabase}
\usage{
metabase_query2(sql_query, database_id = Sys.getenv("METABASE_DATABASE_ID"))
}
\arguments{
\item{sql_query}{SQL query to execute}

\item{database_id}{Database ID to query, will be set by \code{\link{metabase_setup}}.}
}
\value{
data.frame containing the results of the query
}
\description{
Sends an SQL query to Metabase for the given database.
Data is limited to 2000 records by the server using this approach.
This might be faster for small lookups of the DB.
}
