% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabaser.R
\name{metabase_query_sql}
\alias{metabase_query_sql}
\title{Query Metabase}
\usage{
metabase_query_sql(
  sql_query,
  handle = metabase_login_default(),
  col_types = readr::cols(.default = readr::col_character()),
  use_cache = TRUE,
  ...
)
}
\arguments{
\item{sql_query}{SQL query to execute}

\item{handle}{metabase_handle object}

\item{col_types}{default column type can be set to character for all}

\item{use_cache}{if TRUE will use the cache if it is available}
}
\value{
data.frame containing the results of the query
}
\description{
Sends an SQL query to Metabase for the given database.
Data is limited to 2000 records by the server using this approach.
This might be faster for small lookups of the DB.
}
