% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabaser.R
\name{metabase_set_default}
\alias{metabase_set_default}
\title{Sets a default to be used for matabaser}
\usage{
metabase_set_default(param, value)
}
\arguments{
\item{param}{normally one of base_url, database_id, creds_file, username, password}

\item{value}{the value for which ever one is being set}
}
\description{
typically the user will set all values at once from the function metabase_set_defaults
}
