% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabaser.R
\name{metabase_set_defaults}
\alias{metabase_set_defaults}
\title{Sets the defaults to be used for metabaser}
\usage{
metabase_set_defaults(
  base_url = NULL,
  database_id = NULL,
  creds_file = NULL,
  username = NULL,
  pw = NULL
)
}
\arguments{
\item{base_url}{Base URL for the Metabase API}

\item{database_id}{Database ID to connect to}

\item{creds_file}{File containing Metabase account credentials to connect with}

\item{username}{Username}

\item{password}{Password}
}
\description{
Sets the defaults to be used for metabaser
}
