% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabaser.R
\name{metabase_setup}
\alias{metabase_setup}
\title{Setup Metabase connection}
\usage{
metabase_setup(base_url, database_id, creds_file = "~/metabase_creds")
}
\arguments{
\item{base_url}{Base URL for the Metabase API}

\item{database_id}{Database ID to connect to}

\item{creds_file}{File containing Metabase account credentials to connect with}
}
\description{
Sets environment variables for the connection to Metabase.
}
