% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabaser.R
\name{metabase_status}
\alias{metabase_status}
\title{Check Metabase status}
\usage{
metabase_status(base_url = metabase_url())
}
\arguments{
\item{base_url}{Base URL for the Metabase API}
}
\value{
TRUE if a session is active with a logged-in user, FALSE otherwise.
}
\description{
Check Metabase status
}
