% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabaser.R
\name{metabase_url}
\alias{metabase_url}
\title{Get/Set Metabase URL}
\usage{
metabase_url(url = NULL)
}
\description{
If no argument is given, retrieves the current Metabase URL.
If an argument is given, sets the Metabase URL.
}
