% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabaser.R
\name{metabase_login}
\alias{metabase_login}
\title{Login to Metabase}
\usage{
metabase_login(
  base_url,
  database_id,
  creds_file = NULL,
  username = NULL,
  password = NULL,
  auto_setup = TRUE
)
}
\arguments{
\item{base_url}{Base URL for the Metabase API}

\item{database_id}{Database ID to connect to}

\item{creds_file}{File containing Metabase account credentials to connect with}

\item{username}{Username}

\item{password}{Password
if FALSE, requires \code{\link{metabase_setup}} to be executed before}
}
\description{
This will login to the Metabase API with the given username and password.
Credentials must be specified either directly through the \code{username} and \code{password} parameters
or as a file through the \code{creds_file} parameter.
}
\details{
The file should have "username=username" on one line and "password=password" on the next.
If the login is successful, a cookie will be set behind the scenes with a session ID for all future requests.
}
