% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabaser.R
\name{metabase_login}
\alias{metabase_login}
\title{Login to Metabase}
\usage{
metabase_login(creds_file = Sys.getenv("METABASE_CREDS_FILE"))
}
\arguments{
\item{creds_file}{File containing Metabase account credentials to connect with}
}
\description{
This will login to the Metabase API with the username and password specified in the given file.
The file should have "username=username" on one line and "password=password" on the next.
If the login is successful, a cookie will be set behind the scense with a session ID for all future requests.
If \code{\link{metabase_setup}} was used, \code{metabase_login} will be automatically configured with a creds_file.
}
