% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabaser.R
\name{metabase_login}
\alias{metabase_login}
\title{Login to Metabase}
\usage{
metabase_login(
  base_url = NULL,
  database_id = NULL,
  creds_file = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{base_url}{Base URL for the Metabase API}

\item{database_id}{Database ID to connect to}

\item{creds_file}{File containing Metabase account credentials to connect with}

\item{username}{Username}

\item{password}{Password}
}
\description{
This will login to the Metabase API with the given username and password.
Credentials must be specified either directly through the \code{username} and \code{password} parameters
or as a file through the \code{creds_file} parameter.
}
\details{
If used, the creds_file should have "username=username" on one line and "password=password" on the next.
If the login is successful, a \code{metabase_handle} object will be returned to use in subsequent functions.
}
