% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabaser.R
\name{metabase_status}
\alias{metabase_status}
\title{Check Metabase status}
\usage{
metabase_status()
}
\value{
TRUE if a session is active with a logged-in user, FALSE otherwise.
}
\description{
Check Metabase status
}
